requirements.txt
PyQt6>=6.11.0
opencv-python>=4.13.0
numpy>=2.4.0
Pillow>=12.2.0
python-dotenv>=1.2.0
